Manager method to create an image from a URL

+++ bulbs/images/tests.py
@@ -5,8 +5,7 @@ from django.conf import settings
 from django.test.client import Client
 
 from django.test import TestCase
-import requests
-import StringIO
+import os
 import shutil
 
 from bulbs.images.models import Image, ImageAspectRatio
@@ -17,20 +16,51 @@ class ImageTagsTestCase(TestCase):
     def setUp(self):
         for ratio in [(16, 9), (1, 1), (3, 4)]:
             ImageAspectRatio.objects.get_or_create(width=ratio[0], height=ratio[1], slug='%sx%s' % ratio)
-        image = Image.objects.create(_width=300, height=168, alt="test image", caption="test_image")
-        r = requests.get("http://o.onionstatic.com/images/18/18053/original/1200.jpg")
-        image_file = StringIO.StringIO()
-        image_file.write(r.content)
-        image.original.save(
-            "1200.jpg",
-            File(image_file)
+        image = Image.objects.create(
+            _width=1200,
+            _height=769,
+            alt="test image",
+            caption="test_image"
         )
-        image.save()
 
         self.context = {'image': image}
 
     def tearDown(self):
-        shutil.rmtree(settings.MEDIA_ROOT)
+        if os.path.exists(settings.MEDIA_ROOT):
+            shutil.rmtree(settings.MEDIA_ROOT)
+
+    def test_image_creation(self):
+        image = Image.objects.create_from_url(
+            url="http://o.onionstatic.com/images/18/18053/original/1200.jpg",
+            alt="test image",
+            caption="test_image"
+        )
+        self.assertEqual(image.caption, "test_image")
+        self.assertEqual(image.width, 1200)
+        self.assertEqual(image.height, 679)
+        self.assertTrue(os.path.exists(image.original.path))
+
+        client = Client()
+        base_url = '/images/crops/%s/' % image.id
+        response = client.get(base_url + '3x4/100_90.jpg')
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.get('Content-Type'), 'image/jpeg')
+
+        response = client.get(base_url + '3x4/100_100.png')
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.get('Content-Type'), 'image/png')
+
+        response = client.get(base_url + '3x4/100_47.jpg')
+        self.assertEqual(response.status_code, 404)
+
+        response = client.get(base_url + '3x4/100_47.png')
+        self.assertEqual(response.status_code, 404)
+
+        response = client.get(base_url + '7x9/100_90.jpg')
+        self.assertEqual(response.status_code, 404)
+
+        response = client.get('/images/crops/666/3x4/100_90.jpg')
+        self.assertEqual(response.status_code, 404)
 
     def testImageUrlTag(self):
 
@@ -54,25 +84,3 @@ class ImageTagsTestCase(TestCase):
 
         rendered = test_template.render(test_context)
         self.assertEqual(rendered, '<img src="/images/crops/1/3x4/100_90.jpg" alt="test image" />')
-
-    def testImageCropping(self):
-        client = Client()
-        response = client.get('/images/crops/1/3x4/100_90.jpg')
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.get('Content-Type'), 'image/jpeg')
-
-        response = client.get('/images/crops/1/3x4/100_100.png')
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.get('Content-Type'), 'image/png')
-
-        response = client.get('/images/crops/1/3x4/100_47.jpg')
-        self.assertEqual(response.status_code, 404)
-
-        response = client.get('/images/crops/1/3x4/100_47.png')
-        self.assertEqual(response.status_code, 404)
-
-        response = client.get('/images/crops/1/7x9/100_90.jpg')
-        self.assertEqual(response.status_code, 404)
-
-        response = client.get('/images/crops/2/3x4/100_90.jpg')
-        self.assertEqual(response.status_code, 404)
